Add files via upload
task 1 
This webpage is created using basic HTML structure with a heading and two paragraphs. CSS is applied inside the <style> tag to centre the heading and set the paragraph colour to dark grey. The page demonstrates basic formatting and styling concepts used in web development.

task 7
This webpage displays an <h2> heading and a button. Using JavaScript’s getElementById and innerHTML, the text of the heading is changed from “Before Click” to “After Click” when the button is pressed. This demonstrates basic DOM manipulation